+++ .prettierrc.js
@@ -1,0 +1,7 @@
+export default {
+  useTabs: false,
+  tabWidth: 2,
+  semi: false,
+  trailingComma: "none",
+  endOfLine: "lf"
+}