+++ client/components/BaseNavbar.vue
@@ -5,10 +5,9 @@ const navLinks = [
     subitems: [
       { name: "By species", link: "/browse?filter=species" },
       { name: "By character", link: "/browse?filter=characters" },
-      { name: "By chronological order", link: "/browse?filter=chrono" }
+      { name: "By chronological order", link: "/browse?filter=chronological" }
     ]
   },
-  { name: "Reports" },
   { name: "About" }
 ]
 </script>