+++ client/pages/about/index.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="prose-headings:font-heading">
+    <ContentDoc />
+  </div>
+</template>