+++ python/logger.py
@@ -1,10 +1,8 @@
 """
 ## Logger
 
-It logs things. Pretty neat, I guess lol
-
-Copyright 2021-2023 Kerby Keith Aquino
-MIT License
+Copyright 2021-2024 Kerby Keith Aquino
+Licensed under Apache-2.0
 """
 import logging
 from sys import platform